Changed the torch_dtype from torch.float32 to torch.bloat16 in case of CPU for efficiency

+++ node-hub/dora-phi4/dora_phi4/main.py
@@ -23,7 +23,7 @@ elif torch.backends.mps.is_available():
     torch_dtype = torch.float16  # Reduce memory usage for MPS
 else:
     device = "cpu"
-    torch_dtype = torch.float32  # CPU uses float32
+    torch_dtype = torch.bfloat16  # CPU uses bfloat16 for efficiency
 
 
 # Load the model and processor
